+++ web/js/annotation_sidebar/child_annotations/comments/ViewComment.tsx
@@ -47,7 +47,7 @@ export class ViewComment extends React.Component<IProps, IState> {
 
             <div className="m-1 mb-2">
 
-                <div key={key} className="comment">
+                <div key={key} className="comment muted-color-root">
 
                     <div className="pb-1 pt-1">